Added alex's button, and removed graveyard. Also got rid of nested .maps()'s :(
